// src/app/features/productos/models/producto.model.ts
export interface Producto {
  id: number;
  nombre: string;
  precio: number;
  stock: number;
  estado: boolean;
  categoriaId: number;
  categoriaNombre: string;
  fechaCreacion: string;
  fechaModificacion: string | null;
}

export interface ProductoRequest {
  nombre: string;
  precio: number;
  stock: number;
  estado: boolean;
  categoriaId: number;
}

export type OrdenProducto = 'id' | 'nombre' | 'precio' | 'stock';

export type Direccion = 'asc' | 'desc';